requirements.txt
MetaTrader5
pandas
numpy
requests
urllib3
python-dotenv
yfinance
beautifulsoup4
